+++ arduino/src/BlockSelector.cpp
@@ -2,7 +2,7 @@
 #include "Selector.h"
 #include "BlockSelector.h"
 
-BlockSelector::BlockSelector(int minR, int maxR, int minC, int maxC)
+BlockSelector::BlockSelector(byte minR, byte maxR, byte minC, byte maxC)
   : Selector()
 {
   minRow = minR;
@@ -11,24 +11,24 @@ BlockSelector::BlockSelector(int minR, int maxR, int minC, int maxC)
   maxCol = maxC;
 }
 
-unsigned long BlockSelector::getPinTime(unsigned long currentTime, int pin)
+unsigned long BlockSelector::getPinTime(unsigned long currentTime, byte pin)
 {
   currentRow = Row(pin);
   currentCol = Col(pin);
   if (currentRow >= minRow && currentRow <= maxRow && currentCol >= minCol && currentCol <= maxCol) {
-    //Serial.println(String(pin) + " row: " + String(currentRow) + " col: " + String(currentCol) + " included");
+    //Serial.prbyteln(String(pin) + " row: " + String(currentRow) + " col: " + String(currentCol) + " included");
     return currentTime;
   }
-    //Serial.println(String(pin) + " row: " + String(currentRow) + " col: " + String(currentCol) + " not included");
+    //Serial.prbyteln(String(pin) + " row: " + String(currentRow) + " col: " + String(currentCol) + " not included");
   return -1;
 }
 
-int BlockSelector::Row(int pin)
+byte BlockSelector::Row(byte pin)
 {
   return pin / STRIP_COLS;
 }
 
-int BlockSelector::Col(int pin)
+byte BlockSelector::Col(byte pin)
 {
   return pin % STRIP_COLS;
 }
