+++ arduino/src/Selector.cpp
@@ -5,7 +5,7 @@ Selector::Selector()
 {
 }
 
-unsigned long Selector::getPinTime(unsigned long currentTime, int pin)
+unsigned long Selector::getPinTime(unsigned long currentTime, byte pin)
 {
   return currentTime;
 }
